Move paid ask mode to Trigger.dev
