+++ trigger/paid-ask.ts
@@ -1,0 +1,6 @@
+import { createChatTriggerTask } from "./chat-task";
+
+export const paidAskTask = createChatTriggerTask({
+  id: "paid-ask",
+  defaultMode: "ask",
+});
